utils/network: move network utilities to utils.network
configure_network.py is now a more complete module and has more than
just "configure methods". Also we have another module called
"network.py" with some port related code. This change will introduce the
"network package" inside utils. This layout it is a proposal for a
better organization.
